give the oracle the key
since it’s apparently impossible to create a global vec with the 16 byte
key that’s supposed to stay the same every time the oracle is called,
I’m going to have to fake it by generating the key and then passing it
in to the oracle every time I call it. blah.